+++ src/Brokering/Winterhaven.Brokering/Events/Users/UserLoggedOutEvent.cs
@@ -3,14 +3,37 @@
 namespace Winterhaven.Brokering.Events.Users;
 
 /// <summary>
-///   Represents an event that is triggered when a user successfully logs out.
 /// </summary>
-/// <param name="Identifier">
-///   The unique identifier of the user who has logged out.
-/// </param>
-/// <param name="AccessToken">
-///   The access token of the user who has logged out.
-/// </param>
-public sealed record UserLoggedOutEvent(
-    Guid Identifier,
-    string AccessToken);
+public sealed record UserLoggedOutEvent : IEvent
+{
+    private const string EventRoute = "user.logged_out";
+
+    /// <summary>
+    /// </summary>
+    public UserLoggedOutEvent(Guid userAccountId, string accessToken)
+    {
+        if (userAccountId == Guid.Empty)
+        {
+            throw new ArgumentException($"{userAccountId} must not be empty.", nameof(userAccountId));
+        }
+
+        ArgumentException.ThrowIfNullOrWhiteSpace(accessToken);
+
+        UserAccountId = userAccountId;
+        AccessToken = accessToken;
+    }
+
+    /// <summary>
+    /// </summary>
+    public string AccessToken { get; init; }
+
+    /// <summary>
+    /// </summary>
+    public Guid UserAccountId { get; init; }
+
+    /// <inheritdoc/>
+    public static string GetPublishEventRoute(PublishOptions options) => EventRoute;
+
+    /// <inheritdoc/>
+    public static string GetSubscribeEventRoute(SubscribeOptions options) => EventRoute;
+}