+++ src/Gateway/Winterhaven.Gateway.Presentation/Targets/Players/PlayerRpcTarget.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using StreamJsonRpc;
+using Winterhaven.Brokering;
+using Winterhaven.Brokering.Events.Players;
+using Winterhaven.Common.DTOs.Players;
+using Winterhaven.Gateway.Core.Application.Services.Users;
+using Winterhaven.Gateway.Presentation.Attributes;
+
+namespace Winterhaven.Gateway.Presentation.Targets.Players;
+
+internal sealed record PlayerActionRpcParameters(
+    PlayerActionDto[] ActionQueue);
+
+internal sealed class PlayerRpcTarget : IRpcTarget
+{
+    private readonly IMessageBus messageBus;
+
+    private readonly IUserSessionContext userSessionContext;
+
+    public PlayerRpcTarget(IMessageBus messageBus, IUserSessionContext userSessionContext)
+    {
+        this.messageBus = messageBus ?? throw new ArgumentNullException(nameof(messageBus));
+        this.userSessionContext = userSessionContext ?? throw new ArgumentNullException(nameof(userSessionContext));
+    }
+
+    [JsonRpcAuthorize]
+    [JsonRpcMethod("player.action", UseSingleObjectParameterDeserialization = true)]
+    public async Task PerformAction(PlayerActionRpcParameters parameters, CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(parameters);
+
+        var notification = new PlayerActionEvent(parameters.ActionQueue);
+
+        var options = new PublishOptions()
+            .WithRouteKey("playerId", userSessionContext.UserSession!.UserAccountId.ToString());
+
+        await messageBus.PublishAsync(notification, options, cancellationToken).ConfigureAwait(false);
+    }
+}